+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/menu/advanced_redstone_interface/AdvancedRedstoneInterfaceMenu.kt
@@ -25,11 +25,10 @@ class AdvancedRedstoneInterfaceMenu(
 			)
 
 	init {
-		addSlots()
-		addPlayerInventorySlots(51)
+		addSlots(51)
 	}
 
-	override fun addSlots() {
+	override fun addContainerSlots() {
 		val y = 18
 
 		for (i in 0 until RedstoneInterfaceAdvancedBlockEntity.CONTAINER_SIZE) {