+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed the Moon Phase Detector desync (#72)
 - Fixed the Spectre Illuminator not working with Sodium
 - Fix images not working in wiki (#75)
+- Fixed Block Detector not letting Redstone wire visually connect to it
 
 ### Changed